# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/categories/{id}": {
      "get": {
        "operationId": "getPublicCategoryById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "PublicCategory"
        ]
      }
    },
    "/api/admin/categories": {
      "get": {
        "operationId": "getAdminCategories",
        "parameters": [
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter by category name",
            "schema": {
              "example": "table",
              "type": "string"
            }
          },
          {
            "name": "description",
            "required": false,
            "in": "query",
            "description": "Filter by category description",
            "schema": {
              "example": "dining",
              "type": "string"
            }
          },
          {
            "name": "parentId",
            "required": false,
            "in": "query",
            "description": "Filter by parent category id",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 25,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "isDeleted",
            "required": false,
            "in": "query",
            "description": "Include soft-deleted categories",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getAdminCategoryByIdFromList": {
                "operationId": "getAdminCategoryById",
                "parameters": {
                  "id": "$response.body#/items/0/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryPaginationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "AdminCategory"
        ]
      },
      "post": {
        "operationId": "createAdminCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "links": {
              "getCreatedAdminCategory": {
                "operationId": "getAdminCategoryById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              },
              "updateCreatedAdminCategory": {
                "operationId": "updateAdminCategory",
                "parameters": {
                  "id": "$response.body#/id"
                }
              },
              "deleteCreatedAdminCategory": {
                "operationId": "deleteAdminCategory",
                "parameters": {
                  "id": "$response.body#/id"
                }
              },
              "restoreCreatedAdminCategory": {
                "operationId": "restoreAdminCategory",
                "parameters": {
                  "id": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "AdminCategory"
        ]
      }
    },
    "/api/admin/categories/{id}": {
      "get": {
        "operationId": "getAdminCategoryById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getPublicCategoryByIdFromAdmin": {
                "operationId": "getPublicCategoryById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "AdminCategory"
        ]
      },
      "patch": {
        "operationId": "updateAdminCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "links": {
              "getUpdatedAdminCategory": {
                "operationId": "getAdminCategoryById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "AdminCategory"
        ]
      },
      "delete": {
        "operationId": "deleteAdminCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "AdminCategory"
        ]
      }
    },
    "/api/admin/categories/{id}/restore": {
      "patch": {
        "operationId": "restoreAdminCategory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getRestoredAdminCategory": {
                "operationId": "getAdminCategoryById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "AdminCategory"
        ]
      }
    },
...
  },
  "info": {
    "title": "catalog-service API",
    "description": "REST API + GRPC endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002"
    }
  ],
  "components": {
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total items count",
            "example": 124
          },
          "page": {
            "type": "number",
            "description": "Current page",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Items per page",
            "example": 25
          }
        },
        "required": [
          "total",
          "page",
          "limit"
        ]
      },
...
      "ImageCategoryDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Image URL",
            "example": "https://cdn.example.com/categories/wooden-tables/main.jpg"
          },
          "type": {
            "type": "string",
            "description": "Image type",
            "example": "main"
          }
        },
        "required": [
          "url",
          "type"
        ]
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "name": {
            "type": "string",
            "example": "Wooden Tables"
          },
          "fullName": {
            "type": "string",
            "example": "Dining Tables and Wooden Sets"
          },
          "slug": {
            "type": "string",
            "example": "wooden-tables"
          },
          "description": {
            "type": "string",
            "example": "Tables for kitchen and dining rooms"
          },
          "parentId": {
            "type": "number",
            "example": 3
          },
          "images": {
            "example": [
              {
                "url": "https://cdn.example.com/categories/wooden-tables/main.jpg",
                "type": "main"
              },
              {
                "url": "https://cdn.example.com/categories/wooden-tables/banner.jpg",
                "type": "banner"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageCategoryDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "fullName",
          "slug",
          "description",
          "parentId",
          "images"
        ]
      },
...
      "AdminCategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "name": {
            "type": "string",
            "example": "Wooden Tables"
          },
          "fullName": {
            "type": "string",
            "example": "Dining Tables and Wooden Sets"
          },
          "slug": {
            "type": "string",
            "example": "wooden-tables"
          },
          "description": {
            "type": "string",
            "example": "Tables for kitchen and dining rooms"
          },
          "parentId": {
            "type": "number",
            "example": 3
          },
          "images": {
            "example": [
              {
                "url": "https://cdn.example.com/categories/wooden-tables/main.jpg",
                "type": "main"
              },
              {
                "url": "https://cdn.example.com/categories/wooden-tables/banner.jpg",
                "type": "banner"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageCategoryDto"
            }
          },
          "sortOrder": {
            "type": "number",
            "example": 120
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-10T08:15:20.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-03-10T09:40:03.000Z"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "id",
          "name",
          "fullName",
          "slug",
          "description",
          "parentId",
          "images",
          "sortOrder",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "AdminCategoryPaginationDto": {
        "type": "object",
        "properties": {
          "meta": {
            "description": "Pagination metadata",
            "example": {
              "total": 124,
              "page": 1,
              "limit": 25
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              }
            ]
          },
          "items": {
            "example": [
              {
                "id": 12,
                "name": "Wooden Tables",
                "fullName": "Dining Tables and Wooden Sets",
                "slug": "wooden-tables",
                "description": "Tables for kitchen and dining rooms",
                "parentId": 3,
                "sortOrder": 120,
                "createdAt": "2026-03-10T08:15:20.000Z",
                "updatedAt": "2026-03-10T09:40:03.000Z",
                "deletedAt": null,
                "images": [
                  {
                    "url": "https://cdn.example.com/categories/wooden-tables/main.jpg",
                    "type": "main"
                  }
                ]
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCategoryDto"
            }
          }
        },
        "required": [
          "meta",
          "items"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Wooden Tables"
          },
          "fullName": {
            "type": "string",
            "description": "Full category name",
            "example": "Dining Tables and Wooden Sets"
          },
          "slug": {
            "type": "string",
            "description": "Unique category slug",
            "example": "wooden-tables"
          },
          "description": {
            "type": "string",
            "description": "Category description",
            "example": "Tables for kitchen and dining rooms"
          },
          "parentId": {
            "type": "number",
            "description": "Parent category id (0 means no parent)",
            "example": 3
          },
          "sortOrder": {
            "type": "number",
            "description": "Sorting priority",
            "example": 120
          }
        },
        "required": [
          "name",
          "slug"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Wooden Tables Premium"
          },
          "fullName": {
            "type": "string",
            "description": "Full category name",
            "example": "Premium Dining Tables and Wooden Sets"
          },
          "slug": {
            "type": "string",
            "description": "Unique category slug",
            "example": "wooden-tables-premium"
          },
          "description": {
            "type": "string",
            "description": "Category description",
            "example": "Updated collection of dining tables"
          },
          "parentId": {
            "type": "number",
            "description": "Parent category id (0 means no parent)",
            "example": 4
          },
          "sortOrder": {
            "type": "number",
            "description": "Sorting priority",
            "example": 80
          }
        }
      },
...
    }
  }
}